Move ResponsiveModal to flat structure
