+++ packages/client/src/views/CorrectionForm/IDVerifier.tsx
@@ -23,10 +23,8 @@ import { injectIntl, WrappedComponentProps as IntlShapeProps } from 'react-intl'
 import { constantsMessages, countryMessages } from '@client/i18n/messages'
 import { messages as certificateMessages } from '@client/i18n/messages/views/certificate'
 import { identityNameMapper } from '@client/forms/identity'
-import {
-  ResponsiveModal,
-  LabelValuePair
-} from '@opencrvs/components/lib/interface'
+import { LabelValuePair } from '@opencrvs/components/lib/interface'
+import { ResponsiveModal } from '@opencrvs/components/lib/ResponsiveModal'
 import { formatLongDate } from '@client/utils/date-formatting'
 import { Content } from '@opencrvs/components/lib/interface/Content'
