+++ packages/client/src/views/PrintCertificate/IDVerifier.tsx
@@ -24,10 +24,8 @@ import { injectIntl, WrappedComponentProps as IntlShapeProps } from 'react-intl'
 import { constantsMessages, countryMessages } from '@client/i18n/messages'
 import { messages as certificateMessages } from '@client/i18n/messages/views/certificate'
 import { identityNameMapper } from '@client/forms/identity'
-import {
-  ResponsiveModal,
-  LabelValuePair
-} from '@opencrvs/components/lib/interface'
+import { LabelValuePair } from '@opencrvs/components/lib/interface'
+import { ResponsiveModal } from '@opencrvs/components/lib/ResponsiveModal'
 import { formatLongDate } from '@client/utils/date-formatting'
 
 interface IVerifierActionProps {
