+++ packages/client/src/views/PrintCertificate/collectorForm/CollectorForm.tsx
@@ -13,10 +13,8 @@ import { PrimaryButton, TertiaryButton } from '@opencrvs/components/lib/buttons'
 import { ErrorText } from '@opencrvs/components/lib/ErrorText'
 import { Content } from '@opencrvs/components/lib/interface/Content'
 
-import {
-  ActionPageLight,
-  ResponsiveModal
-} from '@opencrvs/components/lib/interface'
+import { ActionPageLight } from '@opencrvs/components/lib/interface'
+import { ResponsiveModal } from '@opencrvs/components/lib/ResponsiveModal'
 import {
   createReviewDeclaration,
   modifyDeclaration,
