+++ packages/client/src/views/SysAdmin/Performance/PerformanceHome.tsx
@@ -15,11 +15,8 @@ import { ILocation } from '@client/offline/reducer'
 import { getOfflineData } from '@client/offline/selectors'
 import { IStoreState } from '@client/store'
 import { generateLocations } from '@client/utils/locationUtils'
-import {
-  Box,
-  ISearchLocation,
-  ResponsiveModal
-} from '@opencrvs/components/lib/interface'
+import { Box, ISearchLocation } from '@opencrvs/components/lib/interface'
+import { ResponsiveModal } from '@opencrvs/components/lib/ResponsiveModal'
 import { Spinner } from '@opencrvs/components/lib/Spinner'
 import * as React from 'react'
 import { parse } from 'query-string'
