+++ packages/components/src/interface/index.ts
@@ -29,7 +29,6 @@ export * from './Header/AppHeader'
 export * from './Header/PageHeader'
 export * from './ToggleMenu'
 export * from './ViewData'
-export * from './ResponsiveModal'
 export * from './RadioButton'
 export * from './EventTopBar'
 export * from './ListTable'
